+++ Cargo.toml
@@ -1,8 +1,15 @@
 [package]
-name = "bevy_gltf_components"
+name = "bevy_gltf_flow"
 version = "0.2.0"
 edition = "2021"
 license = "MIT OR Apache-2.0"
+
+[workspace]
+members = [
+    "crates/bevy_gltf_components",
+    "crates/bevy_gltf_blueprints",
+]
+
 
 [dev-dependencies]
 bevy="0.11.2"
@@ -13,6 +20,9 @@ rand = "0.8.5"
 
 [dependencies]
 bevy = { version = "0.11.2", default-features = false, features = ["bevy_asset", "bevy_scene", "bevy_gltf"] }
+bevy_gltf_components = { path = "crates/bevy_gltf_components" }
+bevy_gltf_blueprints = { path = "crates/bevy_gltf_blueprints" }
+
 serde = "*"
 ron="*"
