--- assets/test.ron
@@ -1,3 +1,0 @@
-((
-    blender_worklfow::Player: ()
-))
