+++ crates/bevy_gltf_blueprints/src/clone_entity.rs
