+++ crates/bevy_gltf_components/src/process_gltfs.rs
